% Generated by roxygen2 (4.0.1): do not edit by hand
\name{forest_from_ape}
\alias{ape_from_forest}
\alias{forest_from_ape}
\title{Convert Between Forest and Ape Trees}
\usage{
forest_from_ape(phy)

ape_from_forest(tree)
}
\arguments{
\item{phy}{A tree in ape's "phylo" class.}

\item{tree}{A tree from forest}
}
\description{
Converts between ape trees (objects of class \code{phylo}) and
forest trees (currently \code{xtree} objects composed of
\code{xnode}s.
}
\details{
These functions are lossy: converting from ape to forest and back
may not preserve edge index numbers, though it will preseve names
(and topology, etc).  Converting from forest to ape will lose all
data associated with the tree (that is, the \code{data} field
associated with each node), and only the labels give a route to
re-associating.  Both these may change in future.
}

