\name{aspect_ratio}
\alias{aspect_ratio}
\title{Compute Aspect Ratio}
\usage{
aspect_ratio(object, ...)
}
\arguments{
  \item{object}{Object to compute as}

  \item{...}{Additional arguments to methods, all ignored}
}
\description{
Compute aspect ratio of images (and potentially/eventually
other graphical objects).
}
\details{
This is how much \emph{wider} a thing is than its height.
So an object with width "w" and height "h" has aspect ratio
of w/h,or (w/h):1.  If it is wider than it is high, then
the aspect ratio is greater than one.
}
\author{
Rich FitzJohn
}

