% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanfp}
\alias{cleanfp}
\alias{print.cleanfp}
\alias{cleanfp.default}
\alias{cleanfp.coxph}
\alias{cleanfp.crr}
\alias{cleanfp.crr2}
\alias{cleanfp.coxphf}
\alias{cleanfp.logistf}
\alias{cleanfp.glm}
\alias{cleanfp.table}
\alias{cleanfp.formula}
\title{Clean objects for forest}
\usage{
cleanfp(x, ...)

\method{print}{cleanfp}(x, ...)

\method{cleanfp}{default}(x, ...)

\method{cleanfp}{coxph}(x, exp = TRUE, conf.int = 0.95, digits = 2L, format_pval = TRUE, ...)

\method{cleanfp}{crr}(
  x,
  formula,
  data,
  exp = TRUE,
  conf.int = 0.95,
  digits = 2L,
  format_pval = TRUE,
  ...
)

\method{cleanfp}{crr2}(
  x,
  which = 1L,
  exp = TRUE,
  conf.int = 0.95,
  digits = 2L,
  format_pval = TRUE,
  ...
)

\method{cleanfp}{coxphf}(
  x,
  formula = x$call$formula,
  data,
  exp = TRUE,
  conf.int = 1 - x$alpha,
  digits = 2L,
  format_pval = TRUE,
  ...
)

\method{cleanfp}{logistf}(
  x,
  formula = x$call$formula,
  data,
  exp = TRUE,
  conf.int = 1 - x$alpha,
  digits = 2L,
  format_pval = TRUE,
  ...
)

\method{cleanfp}{glm}(x, exp = TRUE, conf.int = 0.95, digits = 2L, format_pval = TRUE, ...)

\method{cleanfp}{table}(x, conf.int = 0.95, digits = 2L, format_pval = TRUE, ...)

\method{cleanfp}{formula}(
  formula = formula(data),
  data,
  conf.int = 0.95,
  digits = 2L,
  format_pval = TRUE,
  order = c("p.value", "coef", "none"),
  decreasing = FALSE,
  ...
)
}
\arguments{
\item{x}{an object or formula}

\item{...}{additional arguments passed to or from other methods}

\item{exp}{logical; if \code{TRUE}, estimates will be exponentiated}

\item{conf.int}{the confidence level}

\item{digits}{the number of digits past the decimal to keep}

\item{format_pval}{logical or a function used to format p-values to
character strings}

\item{formula}{a formula}

\item{data}{a data frame with variables in \code{formula} and/or
used to fit \code{x}}

\item{which}{for \code{\link[cmprsk2]{crr2}} objects, the index of \code{x}
that will be plotted (default is \code{x[[1]]})}

\item{order}{variable used to order the results, one of \code{"p.value"},
\code{"coef"}, or \code{"none"} for no re-ordering}

\item{decreasing}{logical; if \code{TRUE}, results are sorted by
\code{order} in decreasing order}
}
\description{
Methods to handle cleaning of objects supported by \code{forest}.
}
