\name{forest-internal}
\alias{itree}
\alias{xnode}
\alias{xtree}
\title{Internal Forest Functions}
\description{
  OK, this is currently a cop out until I (a) work out how to document
  Rcpp module-exported classes (which are sort of reference classes
  which are sort of S4 classes) and (b) if these constructors will be
  too low level to export anyway.
}
\keyword{internal}
