% Generated by roxygen2 (4.0.1): do not edit by hand
\name{forest_node}
\alias{forest_node}
\title{Build a Forest Node}
\usage{
forest_node(label = "", length = NA_real_, data = NULL)
}
\arguments{
\item{label}{Label for the node (single-element character vector)}

\item{length}{Branch length from root to tip (single-element
numeric, non-negative)}

\item{data}{Arbitrary data to associate with the node.}
}
\value{
An object of type \code{forest_node}
}
\description{
Build a node.  This is a low-level function that should not be
needed that often.
}
\details{
Nodes are very lightly wrapped lists with the S3 class
\code{forest_node}.  I don't think that we need or want reference
semantics here and this keeps things nice and lightweight.
Methods have been defined for the subset and assignment operators
so that the object only has the elements \code{label},
\code{value} and \code{data}.  This will impose a minor lookup
cost for the assignment operators, but reading should not be
affected except for \code{[} (so prefer \code{\$} and \code{[[}).
}
\author{
Rich FitzJohn
}

