% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forest_utils}
\alias{forest_utils}
\alias{add_reference}
\alias{merge_forest}
\alias{prepare_forest}
\title{Forest utilities}
\usage{
add_reference(
  x,
  header = FALSE,
  total = NULL,
  keep_strata = FALSE,
  keep_cluster = FALSE
)

merge_forest(x, y)

prepare_forest(x)
}
\arguments{
\item{x, y}{objects}

\item{header}{a character vector of header labels for each variable in
the model}

\item{total}{optional total sample size, useful if model excludes
observations with missing data}

\item{keep_strata, keep_cluster}{logical; if \code{FALSE} (default), strata
and/or cluster variables, e.g., \code{y ~ strata(a) + cluster(b)} will be
omitted from row output}
}
\description{
Not recommended to be called by the user.
}
