% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panel_box}
\alias{panel_box}
\title{Panel \code{boxplot}}
\usage{
panel_box(data, y = seq.int(length(data)), col = 1L, axes = FALSE,
  logx = FALSE, add = TRUE, panel.first = NULL, panel.last = NULL,
  ...)
}
\arguments{
\item{data}{a list or data frame for each boxplot}

\item{y}{y-coordinates where each row of \code{data} is plotted}

\item{col}{color for each, recycled as needed}

\item{axes}{logical; if \code{TRUE}, the x-axis is drawn}

\item{logx}{logical; if \code{TRUE}, use a logarithmic x-axis}

\item{add}{logical; if \code{TRUE}, adds to an existing plot; otherwise, a
new plot is created first}

\item{panel.first, panel.last}{expressions to be evaluated before and
after any plotting; see \code{\link{plot.default}}}

\item{...}{additional arguments passed to \code{\link{boxplot}}}
}
\description{
Panel with \code{\link{boxplot}}s.
}
\examples{
set.seed(1)
l <- replicate(5, rnorm(10), simplify = FALSE)

panel_box(l, add = FALSE, axes = FALSE)

panel_box(l, add = FALSE, axes = FALSE,
  panel.first = rect(-1, 0, 1, 6, col = 'grey90', border = NA),
  panel.last = abline(v = 0, lwd = 2, col = 'red')
)

}
\seealso{
Other panel functions: \code{\link{panel_ci}},
  \code{\link{panel_tplot}}
}
\concept{panel functions}
