% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panel_ci}
\alias{panel_ci}
\title{Panel CI}
\usage{
panel_ci(data, y = rev(seq.int(nrow(data))), col = 1L,
  xlim = extendrange(unlist(data)), limits = xlim, logx = FALSE,
  panel.first = NULL, panel.last = NULL, vline = 1, type, ...)
}
\arguments{
\item{data}{a data frame or matrix with three columns containing the
point value, lower limit, and upper limit, respectively}

\item{y}{y-coordinates where each row of \code{data} is plotted}

\item{col}{color for each, recycled as needed}

\item{xlim}{x-axis limits}

\item{limits}{limits for intervals; values outside of \code{limits}
will be truncated and drawn as an arrow}

\item{logx}{logical; if \code{TRUE}, use a logarithmic x-axis}

\item{panel.first, panel.last}{expressions to be evaluated before and
after any plotting; see \code{\link{plot.default}}}

\item{vline}{x-position of the vertical dashed line; use \code{NA} for
no line}

\item{type}{ignored}

\item{...}{additional arguments passed to \code{\link{plot.default}}}
}
\description{
Panel with points and confidence intervals.
}
\examples{
dd <- data.frame(
  x = 1,
  lci = 1 - 1 / 1:10,
  uci = 1 + 1 / 1:10
)

panel_ci(dd[1:10, ])
panel_ci(dd[1:10, ], limits = c(0.25, 1.5))
panel_ci(dd[1:10, ], limits = c(0.25, 0.95))

}
\seealso{
Other panel functions: \code{\link{panel_box}},
  \code{\link{panel_tplot}}
}
