% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panel_tplot}
\alias{panel_tplot}
\title{Panel \code{tplot}}
\usage{
panel_tplot(
  data,
  y = seq.int(length(data)),
  col = 1L,
  axes = FALSE,
  logx = FALSE,
  add = TRUE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{data}{a list or data frame for each tplot}

\item{y}{y-coordinates where each row of \code{data} is plotted}

\item{col}{color for each, recycled as needed}

\item{axes}{logical; if \code{TRUE}, the x-axis is drawn}

\item{logx}{logical; if \code{TRUE}, use a logarithmic x-axis}

\item{add}{logical; if \code{TRUE}, adds to an existing plot; otherwise, a
new plot is created first}

\item{panel.first, panel.last}{expressions to be evaluated before and
after any plotting; see \code{\link{plot.default}}}

\item{...}{additional arguments passed to \code{\link{boxplot}}}
}
\description{
Panel with \code{\link[rawr]{tplot}}s.
}
\examples{
library('rawr')
set.seed(1)
l <- replicate(5, rnorm(10), simplify = FALSE)

panel_tplot(l, add = FALSE)

panel_tplot(l, type = 'd', col = 1:5, pch = 19, add = FALSE,
  panel.first = {
    p <- par('usr')
    rect(-1, p[3L], 1, p[4L], col = 'grey90', border = NA)
  },
  panel.last = abline(v = 0, lwd = 2, col = 'red')
)

}
\seealso{
\code{\link[rawr]{tplot}}

Other panel functions: 
\code{\link{panel_box}()},
\code{\link{panel_ci}()}
}
\concept{panel functions}
