% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.forest}
\alias{summary.forest}
\title{\code{forest} summary}
\usage{
\method{summary}{forest}(object, html = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{\link{forest}}}

\item{html}{logical; if \code{TRUE}, summary will be formatted for display
output in html; see examples}

\item{...}{ignored}
}
\description{
Print a summary of a \code{forest} object.
}
\examples{
x <- forest(
  glm(vs ~ factor(gear) + wt + hp, mtcars, family = 'binomial'),
  header = c('Gear', 'Weight', 'Horsepower')
)
summary(x, html = FALSE)

\dontrun{
library('htmlTable')
s <- summary(x)
s <- gsub('factor.*\\\\)', '', s)
htmlTable(s, align = 'lc', caption = 'Model summary.')
}

}
